requirements.txt
fastapi==0.115.0
uvicorn==0.30.0
huggingface-hub==0.25.0
httpx==0.27.0
python-multipart==0.0.9
pytesseract==0.3.13
Pillow==10.4.0
asyncpg==0.30.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.9.0
pydantic-settings==2.5.0
python-dotenv==1.0.1
PyMuPDF==1.24.0
google-cloud-storage==2.18.0
google-auth==2.35.0
